PrivateTask

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTaskFilter } from './dto/get-task-filter.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task-entity';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,12 +15,12 @@ export class TasksService {
     ) {}
 
 
-   async getAllTasks(filterDto:GetTaskFilter):Promise<Task[]> {
-      return  this.taskrepository.getTasks(filterDto) 
+   async getAllTasks(filterDto:GetTaskFilter,user:User):Promise<Task[]> {
+      return  this.taskrepository.getTasks(filterDto,user) 
     }
 
-   async getTaskById(id:number):Promise<Task> {
-        const found = await this.taskrepository.findOne(id)
+   async getTaskById(id:number,user:User):Promise<Task> {
+        const found = await this.taskrepository.findOne({ where: {id, userId:user.id}})
 
         if(!found) {
             throw new NotFoundException(`Task with ${id} not found`)
@@ -27,19 +28,19 @@ export class TasksService {
         return found
     }
 
-    async createTask(createTaskDto:CreateTaskDto):Promise<Task> {
-        return this.taskrepository.createTask(createTaskDto)
+    async createTask(createTaskDto:CreateTaskDto,user:User):Promise<Task> {
+        return this.taskrepository.createTask(createTaskDto,user)
     }
 
-    async deleteTaskId(id:number):Promise<void> {
-        const found = await this.taskrepository.delete(id)
+    async deleteTaskId(id:number,user:User):Promise<void> {
+        const found = await this.taskrepository.delete({ id, userId: user.id})
         if(found.affected == 0) {
             throw new NotFoundException(`Task with ${id} not found`)
         }
     }
 
-    async updateStatusById(id:number,status:TaskStatus):Promise<Task> {
-        const task =await this.getTaskById(id)
+    async updateStatusById(id:number,status:TaskStatus,user:User):Promise<Task> {
+        const task =await this.getTaskById(id,user)
         task.status = status
         await task.save()
         return task
